Design change: use a dedicated type for conversion & list
Instead of putting everything in the `TimePoint`, this made the
timepoint cumbersome to use in the numerous cases when we did not needs
those functionnalities.
