+++ mithril-aggregator/src/runtime/state_machine.rs
@@ -234,8 +234,13 @@ impl AggregatorRuntime {
                     .map(|om| om.is_expired)
                     .unwrap_or(false);
                 let exists_newer_open_message = {
-                    let new_signed_entity_type =
-                        last_time_point.to_signed_entity(&state.open_message.signed_entity_type);
+                    let new_signed_entity_type = self
+                        .config
+                        .signed_entity_config
+                        .time_point_to_signed_entity(
+                            &state.open_message.signed_entity_type,
+                            &last_time_point,
+                        );
                     new_signed_entity_type != state.open_message.signed_entity_type
                 };
 
@@ -403,7 +408,7 @@ mod tests {
     use mockall::predicate;
     use std::time::Duration;
 
-    use mithril_common::entities::{CardanoTransactionsSigningConfig, Epoch, SignedEntityType};
+    use mithril_common::entities::{Epoch, SignedEntityConfig, SignedEntityType};
     use mithril_common::test_utils::fake_data;
 
     use super::super::runner::MockAggregatorRunner;
@@ -417,7 +422,7 @@ mod tests {
             AggregatorConfig::new(
                 Duration::from_millis(20),
                 fake_data::network(),
-                CardanoTransactionsSigningConfig::dummy(),
+                SignedEntityConfig::dummy(),
             ),
             init_state,
             Arc::new(runner),
