+++ mithril-aggregator/tests/create_certificate.rs
@@ -3,8 +3,8 @@ mod test_extensions;
 use mithril_aggregator::Configuration;
 use mithril_common::{
     entities::{
-        CardanoDbBeacon, ChainPoint, Epoch, ProtocolParameters, SignedEntityConversionConfig,
-        SignedEntityType, SignedEntityTypeDiscriminants, StakeDistributionParty, TimePoint,
+        CardanoDbBeacon, ChainPoint, Epoch, ProtocolParameters, SignedEntityType,
+        SignedEntityTypeDiscriminants, StakeDistributionParty, TimePoint,
     },
     test_utils::MithrilFixtureBuilder,
 };
@@ -23,12 +23,7 @@ async fn create_certificate() {
         ..Configuration::new_sample()
     };
     let mut tester = RuntimeTester::build(
-        TimePoint::new(
-            1,
-            1,
-            ChainPoint::new(10, 1, "block_hash-1"),
-            SignedEntityConversionConfig::dummy(),
-        ),
+        TimePoint::new(1, 1, ChainPoint::new(10, 1, "block_hash-1")),
         configuration,
     )
     .await;
