+++ mithril-aggregator/tests/open_message_expiration.rs
@@ -5,8 +5,8 @@ use std::time::Duration;
 use mithril_aggregator::Configuration;
 use mithril_common::{
     entities::{
-        CardanoDbBeacon, ChainPoint, ProtocolParameters, SignedEntityConversionConfig,
-        SignedEntityType, SignedEntityTypeDiscriminants, TimePoint,
+        CardanoDbBeacon, ChainPoint, ProtocolParameters, SignedEntityType,
+        SignedEntityTypeDiscriminants, TimePoint,
     },
     test_utils::MithrilFixtureBuilder,
 };
@@ -25,12 +25,7 @@ async fn open_message_expiration() {
         ..Configuration::new_sample()
     };
     let mut tester = RuntimeTester::build(
-        TimePoint::new(
-            1,
-            1,
-            ChainPoint::new(10, 1, "block_hash-1"),
-            SignedEntityConversionConfig::dummy(),
-        ),
+        TimePoint::new(1, 1, ChainPoint::new(10, 1, "block_hash-1")),
         configuration,
     )
     .await;
