+++ mithril-common/src/entities/mod.rs
@@ -17,6 +17,7 @@ mod mithril_stake_distribution;
 mod protocol_message;
 mod protocol_parameters;
 mod signed_entity;
+mod signed_entity_config;
 mod signed_entity_type;
 mod signer;
 mod single_signatures;
@@ -41,6 +42,7 @@ pub use mithril_stake_distribution::MithrilStakeDistribution;
 pub use protocol_message::{ProtocolMessage, ProtocolMessagePartKey, ProtocolMessagePartValue};
 pub use protocol_parameters::ProtocolParameters;
 pub use signed_entity::*;
+pub use signed_entity_config::*;
 pub use signed_entity_type::*;
 pub use signer::{Signer, SignerWithStake};
 pub use single_signatures::*;
